Move types into their own file; handle 402; improve comments
